Add files via upload
YourCause-WireFrame
Current Problem

There are numerous charity organization websites, however, there isn’t an all-in-one platform for people to donate or volunteer their time. Many of these organizations are nonprofit so have websites with limited content and technical capabilities. Donors also don’t know if their money is being used in the right places Competitor websites don’t provide the unique aspect of a consolidated place of finding volunteer opportunities while they are also donating.

Purpose

The goal is to create an overall webpage to provide aid to charity organizations. Since many of these organizations are nonprofit, they often use separate websites like GoFundMe to collect donations. Our idea is to create a webpage that is a one-stop for donating to charity organizations as well as provide trust that donations are being sent to the right people.

Target Audience

Charity organizations within Texas are the target audience for this website. We specifically want to choose 2-3 charity organizations to limit the scope of testing the website prototype. The first organization we chose is the North Texas Food Bank. Based on the website statistics, they have provided 96.9 million meals to kids, seniors, and families in need, and more than 34 million pounds of fresh produce were delivered to those that need it. The second organization we chose is Rebuilding Together North Texas. They provide aid to those that need a home repair but cannot afford it and help out local communities in low-income areas. Based on the website statistics, 100,000 people volunteer and help rebuild 10,000 houses a year. The other side of this transaction and also part of the target audience is the donors. They need a place that has a consolidated view of all organizations, to make sure their transaction is secure, and also find volunteer opportunities.